requirements.txt
attrs==23.1.0
autobahn==19.11.2
Automat==22.10.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.5
Flask==3.0.0
hyperlink==21.0.0
idna==3.5
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiteconnect==5.0.0
MarkupSafe==2.1.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zope.interface==6.1
